Make afl scraper only fetch fixture daily still gets games
The fixture is a guide of what games are a list of games that are coming up,
being played, or played. We only need this once per day as this is the guide
for us to know when to query for the more specific match data for a particular
game.